Add files via upload
To run the exe, you need to use the command feature_extration_Content_Independent.exe <filename> in command prompt. You can use batch to do the same repetitively. In the generated feature file, the last two columns are the index of the writer and then the set number of the writer.